+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - `FoldCompound`: new properties `max_bp_span` and `window_size` that
   can be set in constructor
 
+- added `mfe_window()` to calculate mfe substructures for a
+  sliding window
+
 - added `ViennaRNA.init_rand_seed(seedval)` to seed the random
   number generator used by ViennaRNA